requirements.txt
pocketflow>=0.1.0
openai>=1.0.0
youtube-transcript-api>=0.6.0
python-dotenv>=0.21.0
pyyaml>=6.0